+++ modules/home/cli/git.nix
@@ -3,6 +3,7 @@
 in {
   programs.git = {
     enable = true;
+    signing.format = null;
 
     settings = {
       user = {