+++ modules/home/editors/nixvim.nix
@@ -56,7 +56,7 @@
       lualine = {
         enable = true;
         settings = {
-          options = {theme = "catppuccin";};
+          options = {theme = "auto";};
         };
       };
       bufferline.enable = true;