+++ modules/home/gtk.nix
@@ -1,5 +1,6 @@
-{pkgs, ...}: {
+{config, pkgs, ...}: {
   gtk = {
+    gtk4.theme = null;
     iconTheme = {
       name = "Papirus-Dark";
       package = pkgs.papirus-icon-theme;